// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  agentHelpers: typeof agentHelpers;
  agentQueries: typeof agentQueries;
  agents: typeof agents;
  auth: typeof auth;
  chatAgent: typeof chatAgent;
  chats: typeof chats;
  crons: typeof crons;
  memories: typeof memories;
  messages: typeof messages;
  reminderActions: typeof reminderActions;
  reminders: typeof reminders;
  tasks: typeof tasks;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;